Latte: handle multiple (nested) suites
In order to split tests between instances, we need those instances to be able to run arbitrary sets of tests.

Instead of providing a suite name, Latte now accepts a `selection` function that is given the tree of suites, and a flat list of every test. This should let Zen code divvy things up as it sees fit.

The other major part is handing before/after callbacks as we traverse up and down describe trees.