First pass at workers
This introduces the notion of `grep` (which tests to run) and `focus` (which test to debug in head).

Whenever you change `grep`, we send it via websocket to all workers. Workers divide the tests up evenly (by count, not by time) and start running. As tests pass or fail, workers report back via websocket. This is cached on the server, and passed to every connected head.

There's a simple UI to show failing tests. I chose svelte because using a framework with a runtime felt wrong when we're mostly focused on running user code.