Group related failures
I noticed that when many tests fail, there are often only a few underlying reasons. This attempts to group by those reasons, giving a better idea of how much is actually broken.

For any group, clicking on it will show the actual tests that failed, and you can focus any one of them to debug.

I also wired up the "focus next" button, which jumps to the first test of the next failing group.