Some callback failures are fatal
If a before, after, or afterEach fails, the page is in a broken state.

This just reloads the page, but it's a bit lame. If a before filter is permanently broken, it will loop forever. We also do a poor job of communicating the failure to the foreground.

This also splits out afterEach to it's own try/catch block. Even if the test fails, we should be able to run afterEach and move on.