Ability to re-run a failing subset.
One issue with running the full suite of tests is that you often have to wait a to know if your change fixes the issue.

Instead, each pass should run the subset of tests that failed last time we did a full run. The idea is that once a test fails, it's at risk for failing again in the near future.