Drive test via CDP, rather than websockets
This will make writing the Lambda function much easier, since I won't have to set up a websocket to either a server in aws, or try and proxy it back to the local machine.

I think it should also make the code more robust, since no matter what kind of error happens in the browser, we can detect and recover.

The downside is that we're locked into browsers that support CDP. For now that seems ok.

It also makes the code simpler :)