Run tests in parallel on lambda
We run a few hundred chrome instances with this library:
https://github.com/adieuadieu/serverless-chrome

To avoid your local machine trying to host GBs of data, we upload the entire test environment to an S3 bucket (and enable hosting on it).

This environment won't change much between runs, so we keep track of successful uploads and only re-upload files that have changed locally.

Large files can still be slow -- Superhuman's test webpack bundle is 50MB. In this case, we generate a patch and use a Lambda function to apply it.

Lambda has a limit of 1000 concurrent executions, so we work them in groups of 3.