Show when a failing test starts passing
It a test fails when running in the background, but then passes when you debug it, show that group as passing.

This is a bit leaky. We group failures by the error, and it's possible you only fixed one error in a group of many. Even so, it's pretty satisfying to see things turn green.