Faster syncing to S3 from new devices
Previously, we'd use an on-disk cache to keep track of what files we had already sent to S3, and what changes we need to upload.

This doesn't well on a CI server where the disk is more ephemeral. We'd spend a lot of time (not to mention S3 storage) saving the same files over and over. Instead, we should let clients share files when they're unchanged.

Clients generate a manifest of all files and their hashes. The `sync` lambda function takes that and figures out what files are missing from S3 so we can upload them.

When running tests, we use the manifest again to rewrite generic requests to their version-specific ones in S3.  (/lib/latte.js => lib/latte-1e69e8b91.js).

I experimented with a lot of different approaches to make this work. The key constraint was performance: in the span of 1-2 seconds we want to serve 10GB across 10k requests. S3 is actually pretty good at that. I also tried an in-memory node server, lambda@edge, and the API gateway. They were all slow, or we'd exceed their rate limits.