More caching to speed up S3 syncing
This keeps track of files we have already uploaded, their hash and mtime. If the hash hasn't changed, we know the file is already in S3 and there's no need to check it.
    
If the mtime hasn't changed, we know that hash is also unchanged, and the file is already in S3.