Add Oclif
This allows us to break apart our CLI commands
into separate files.

Adds a `reset` command which removes the `.zen` directory. This is useful
when after clearing out the AWS config in order to set it up from a clean state.

In the near future we want to add an `init` command which will deploy the AWS Lambda
functions and layers and set up the zen config.