Upgrade Svelte to v3
Upgrading from v2 to v3 was a pretty large change. I used the guide at
https://rajasegar.github.io/svelte2vs3/ to help.

One of the biggest changes is how Svelte's store now works. It is a separate
module and no longer behaves as a singleton. We create a singleton and attach it
to `Zen.store`. We cannot put computed values in this store so we make "derived"
stores and attach them to the `Zen.computed` key. This commit also moves several
methods that are called on user interaction to the `actions.js` file.

In my testing I found that all the `on:...` event callbacks need to be wrapped in
a method otherwise they are immediately invoked.

Note: The `get` method imported from `svelte/store` is required to access the state
of a store. This method is attached to `Zen.get`. `store.set` also no longer works as
it completely resets the state of the store. We instead want to merge in a new value
and we do this with `.update(mergeValue(valueToMerge))`